fix(traccc): Fix CUDA 13.4 conversion warnings
This commit fixes a few conversion warnings that CUDA 13.4 has started
flagging.

+++ Traccc/device/common/include/traccc/gbts_seeding/device/impl/gbts_bin_spacepoints.ipp
@@ -86,8 +86,9 @@ TRACCC_HOST_DEVICE inline void gbts_bin_spacepoints(
 
     const detray::geometry::identifier geo_id = measurement.surface_link();
     const unsigned int volume = geo_id.volume();
-    const short begin_or_bin =
-        (volume < payload.volumeMapSize) ? volumeToLayerMap[volume] : SHRT_MAX;
+    const short begin_or_bin = (volume < payload.volumeMapSize)
+                                   ? volumeToLayerMap[volume]
+                                   : static_cast<short>(SHRT_MAX);
 
     if (begin_or_bin == SHRT_MAX) {
       reducedSP[globalIndex].w = -CHAR_MAX - 1;
